add github action to add branch prefix to PR title
Add a GitHub action to the stable branch to prefix the PR title with the
branch name (i.e., "[non-main PR]").  The intention behind is to make
backports, or commits in general, to non-master branches more obvious.
The target branch of a pull request is not easily visible in the GitHub
web interface, and that can lead to unintentionally merging PRs to stable
branches.

Unfortunately, I did not manage to automatically extract and set the
target branch name as the PR prefix.  The action [1] supports a %branch%
macro but that evaluates to the source branch.  Hence, for the time
being, we have to hard-code the branch in the .yml file and update it
once we can auto-set the target branch name.  I opened an issue asking
for help.

Since the action enforces a non-empty body template, the PR body is also
prefixed with a bold note that the PR is targetting a non-main branch.

[1] https://github.com/tzkhan/pr-update-action

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>